kern: rewrite memory management

+++ kern/inc/mm/heap.h
@@ -2,10 +2,8 @@
 #include <lib/stdint.h>
 #include <lib/stdbool.h>
 
-/* heapalloc: allocate memory from the kernel heap. */
 void *
-kmalloc(size_t size);
+kmalloc(unsigned long size);
 
-/* heapfree: free memory to the kernel heap. */
 void
 kfree(void *ptr);
